Organized project structure: moved unused files to codetest, kept only main files in root, updated verification script
