--- SUBMISSION_SUMMARY.md
@@ -1,72 +1,0 @@
-# Reinforcement Learning Assignment Submission
-
-**Student:** Sanat Popli  
-**Course:** Reinforcement Learning for Agentic AI Systems  
-**Date:** August 10, 2025
-
-## Assignment Completion Summary
-
-### ✅ **Core Requirements Implemented:**
-
-1. **Value-Based Learning (DQN)**
-   - File: `src/rl/dqn_agent.py`
-   - Features: Experience replay, target networks, ε-greedy exploration
-   - Application: Content selection and question sequencing
-
-2. **Policy Gradient Methods (PPO)**
-   - File: `src/rl/ppo_agent.py`
-   - Features: Actor-critic architecture, clipped surrogate objective, GAE
-   - Application: Strategic teaching adaptation
-
-3. **Multi-Agent Coordination**
-   - File: `src/orchestration/tutorial_orchestrator.py`
-   - Strategies: Hierarchical, competitive, collaborative coordination
-   - Implementation: Context-aware agent selection and coordination
-
-### 🚀 **Advanced Features:**
-
-- **Experimental Framework**: `src/experiments/experiment_runner.py`
-- **Student Modeling**: Dynamic profiles with learning progression
-- **Data Analytics**: Comprehensive progress tracking and reporting
-- **Multiple Interfaces**: Demo, training, evaluation, human interaction
-
-### 📁 **Key Files to Review:**
-
-1. **`complete_assignment_demo.py`** - Main demonstration of all features
-2. **`docs/TECHNICAL_REPORT.md`** - Comprehensive technical documentation
-3. **`README.md`** - Project overview and setup instructions
-4. **`src/rl/`** - Core RL algorithm implementations
-5. **`src/experiments/experiment_runner.py`** - Evaluation framework
-
-### 🏃‍♂️ **Quick Start for Evaluation:**
-
-```bash
-# Run main demonstration
-python complete_assignment_demo.py
-
-# Run human interactive session
-python human_interactive_tutor.py
-
-# View generated results
-python view_results_demo.py
-
-# Run experimental evaluation (optional - takes 2-3 hours)
-python src/main.py --mode experiment
-```
-
-### 📊 **Project Highlights:**
-
-- **Advanced Architecture**: Multi-agent coordination with specialized RL algorithms
-- **Educational Application**: Practical tutoring system with real-world relevance
-- **Comprehensive Evaluation**: Statistical analysis and visualization framework
-- **Professional Documentation**: Technical report with implementation details
-- **Data Persistence**: Complete analytics and progress tracking system
-
-### 🎯 **Innovation Points:**
-
-1. **Novel Multi-Agent RL Application** in educational technology
-2. **Hierarchical Coordination Strategy** for competing objectives
-3. **Dynamic Student Modeling** with realistic behavioral simulation
-4. **Comprehensive Experimental Framework** for systematic evaluation
-
-This project demonstrates advanced understanding of reinforcement learning concepts and their practical application in a complex, multi-agent educational system.
