+++ codetest/START_HUMAN_TUTOR.bat
@@ -1,0 +1,12 @@
+@echo off
+echo ====================================================
+echo    HUMAN INTERACTIVE AI TUTORING SYSTEM
+echo    Easy Launch for Windows
+echo ====================================================
+echo.
+
+cd /d "%~dp0"
+
+python launch_human_tutor.py
+
+pause
